+++ rs/moq-mux/src/catalog/mod.rs
@@ -35,4 +35,4 @@ pub use format::*;
 pub use producer::{Guard, Producer};
 pub use select::Select;
 pub use stream::Stream;
-pub use tracks::{AudioTrack, VideoTrack};
+pub use tracks::{AudioTrack, CustomTrack, CustomTrackExt, VideoTrack};